# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/employees": {
      "post": {
        "description": "Creates a new employee record in the system. Requires admin, HR manager, or HR officer role. The employee data is provided in the request body, and the creator information is extracted from the JWT token.",
        "operationId": "EmployeesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEmployeeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new employee",
        "tags": [
          "employees"
        ]
      },
      "get": {
        "description": "Retrieves a paginated list of all employees. Requires admin, HR manager, or HR officer role. Results can be paginated using page and limit query parameters. The user information is extracted from the JWT token to determine access permissions.",
        "operationId": "EmployeesController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all employees",
        "tags": [
          "employees"
        ]
      }
    },
    "/employees/{employeeNumber}": {
      "get": {
        "description": "Retrieves detailed information about a specific employee identified by their employee number. The user information is extracted from the JWT token to determine access permissions. Users can only access their own information or information of employees they have permission to view based on their role.",
        "operationId": "EmployeesController_findOne",
        "parameters": [
          {
            "name": "employeeNumber",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get an employee by employeeNumber",
        "tags": [
          "employees"
        ]
      },
      "patch": {
        "description": "Updates information for a specific employee identified by their employee number. Requires admin, HR manager, or HR officer role. The update data is provided in the request body, and the user information is extracted from the JWT token to determine access permissions and track who made the changes.",
        "operationId": "EmployeesController_update",
        "parameters": [
          {
            "name": "employeeNumber",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEmployeeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update an employee",
        "tags": [
          "employees"
        ]
      },
      "delete": {
        "description": "Removes an employee record from the system. This operation is restricted to administrators only. The employee is identified by their employee number. The user information is extracted from the JWT token to track who performed the deletion. This operation may be soft delete depending on the implementation.",
        "operationId": "EmployeesController_remove",
        "parameters": [
          {
            "name": "employeeNumber",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete an employee",
        "tags": [
          "employees"
        ]
      }
    },
    "/employees/{employeeNumber}/role": {
      "patch": {
        "description": "Assigns a specific role to an employee. This operation is restricted to administrators only. The employee is identified by their employee number, and the new role is provided in the request body. This endpoint allows administrators to promote employees to different roles such as HR_MANAGER, HR_OFFICER, or MANAGER.",
        "operationId": "EmployeesController_assignRole",
        "parameters": [
          {
            "name": "employeeNumber",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Assign a role to an employee",
        "tags": [
          "employees"
        ]
      },
      "delete": {
        "description": "Removes any special role from an employee and sets their role back to the default EMPLOYEE role. This operation is restricted to administrators only. The employee is identified by their employee number. This endpoint allows administrators to demote employees who no longer need elevated permissions.",
        "operationId": "EmployeesController_unassignRole",
        "parameters": [
          {
            "name": "employeeNumber",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Unassign role from an employee",
        "tags": [
          "employees"
        ]
      }
    },
...
    "/performance-reviews/employee/{employeeId}": {
      "get": {
        "description": "Retrieves all performance reviews for a specific employee identified by their ID. Regular employees can only access their own reviews, while HR staff and managers can access reviews for employees they manage. The user information is extracted from the JWT token to determine access permissions. This endpoint is useful for viewing an employee's performance history over time.",
        "operationId": "PerformanceReviewsController_getEmployeeReviews",
        "parameters": [
          {
            "name": "employeeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all reviews for a specific employee",
        "tags": [
          "performance-reviews"
        ]
      }
    },
...
    "/payroll/employee/{employeeId}": {
      "get": {
        "description": "Retrieves payroll records for a specific employee identified by their ID. Results can be filtered by date range using startMonth, startYear, endMonth, and endYear query parameters. Regular employees can only access their own payroll records, while HR staff and managers can access payroll records for employees they manage. The user information is extracted from the JWT token to determine access permissions.",
        "operationId": "PayrollController_getEmployeePayroll",
        "parameters": [
          {
            "name": "employeeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startMonth",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "startYear",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "endMonth",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "endYear",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get employee payroll records",
        "tags": [
          "payroll"
        ]
      }
    },
...
    "/attendance/employee/{employeeId}": {
      "get": {
        "description": "Retrieves attendance records for a specific employee identified by their ID. Results can be filtered by date range using startDate and endDate query parameters. Regular employees can only access their own attendance records, while HR staff and managers can access attendance records for employees they manage. The user information is extracted from the JWT token to determine access permissions.",
        "operationId": "AttendanceController_getEmployeeAttendance",
        "parameters": [
          {
            "name": "employeeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "required": true,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": true,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get employee attendance records",
        "tags": [
          "attendance"
        ]
      }
    },
...
  },
  "info": {
    "title": "HR Management API",
    "description": "API for managing employees and documents",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateEmployeeDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "employeeNumber": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "contractStartDate": {
            "format": "date-time",
            "type": "string"
          },
          "contractEndDate": {
            "format": "date-time",
            "type": "string"
          },
          "contractType": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "position": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "firstName",
          "lastName",
          "email",
          "employeeNumber",
          "phoneNumber",
          "role",
          "contractStartDate",
          "contractType",
          "location",
          "position",
          "password"
        ]
      },
      "UpdateEmployeeDto": {
        "type": "object",
        "properties": {}
      },
      "AssignRoleDto": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "description": "The role to assign to the employee",
            "enum": [
              "ADMIN",
              "HR_MANAGER",
              "HR_OFFICER",
              "MANAGER",
              "EMPLOYEE"
            ],
            "example": "HR_MANAGER"
          }
        },
        "required": [
          "role"
        ]
      },
...
    }
  }
}